+++ docs/sidebars.js
@@ -12,7 +12,11 @@ module.exports = {
       {
         type: 'category',
         label: 'Plugins',
-        items: ['plugins_api/PluginCustomization', 'plugins_api/GiphyPlugin'],
+        items: [
+          'plugins_api/PluginCustomization',
+          'plugins_api/GiphyPlugin',
+          'plugins_api/VideoPlugin',
+        ],
       },
       {
         type: 'category',